add: jwt for bearer token & passlib for hashing
Added JWT for the bearer token and passlib for hashing the passwords to
make sure they match with the password received on the /token endpoint.

Updated the logic to make it more secure by using the new libraries
which also means can now save the hashed password in the db.